test/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
paho-mqtt==1.5.0
pkginfo==1.5.0.1
tqdm==4.30.0
